#75 Refactor : Style폴더 구조 변경
