+++ src/components/main/taveAnimation.js
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
-import * as t from "../../styles/taveAnimationStyle"
 import { SetModal } from "../../redux/actions/mainAction";
 import { useDispatch } from "react-redux";
+import * as t from "../../styles/main/taveAnimationStyle"
 
 const TaveAnimation = ({secondAuthStatus}) => {
   const navigate = useNavigate();
