+++ src/styles/main/loginModalStyle.js
