+++ src/styles/main/mainPageStyle.js
