+++ src/styles/main/mobileStyle.js
