+++ src/styles/main/taveAnimationStyle.js
